fix bug

+++ src/app/pages/staff/staff.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ConfirmModalComponent } from 'app/components/confirm-modal/confirm-modal.component';
 import { STATUS_CODE } from 'app/constant/constant';
 import { IStaffResponse } from 'app/interfaces/serve-response';
+import { HelperService } from 'app/services/helper.service';
 import { BsModalService } from 'ngx-bootstrap/modal';
 import { AddEditStaffComponent } from './components/add-edit-staff/add-edit-staff.component';
 import { StaffService } from './services/staff.service';
@@ -14,7 +16,8 @@ export class StaffComponent implements OnInit {
   staffs: IStaffResponse[] = []
   constructor(
     private staffService: StaffService,
-    private modalService: BsModalService
+    private modalService: BsModalService,
+    private helperService: HelperService
   ) { }
 
   ngOnInit(): void {
@@ -50,10 +53,16 @@ export class StaffComponent implements OnInit {
   }
 
   deleteStaff(id: number, $index: number) {
-    this.staffService.deleteStaff(id).subscribe(res => {
-      if (res.code === STATUS_CODE.SUCCESS && res.data) {
-        this.staffs.splice($index, 1);
+    const modal = this.modalService.show(ConfirmModalComponent);
+    modal.content.onClick.subscribe(res => {
+      if (res) {
+        this.staffService.deleteStaff(id).subscribe(res => {
+          if (res.code === STATUS_CODE.SUCCESS && res.data) {
+            this.staffs.splice($index, 1);
+          }
+        })
       }
+      modal.hide();
     })
   }
 }
